AI_Farmtopia/farming_whiz.py
```python
import random
from ai_components import AIComponentSchema

class FarmingWhiz(AIComponentSchema):
    def __init__(self, playerData, gameSettings):
        super().__init__(playerData, gameSettings)
        self.whiz_tips = [
            "Remember to water your crops regularly!",
            "Different crops grow best in different seasons.",
            "Animals need care and attention to thrive.",
            "Weather can affect your farm's productivity.",
            "Don't forget to rest, farming is hard work!",
            "Use power-ups to boost your progress.",
            "Complete quests for extra rewards."
        ]

    def generate_tip(self):
        tip_index = random.randint(0, len(self.whiz_tips) - 1)
        return self.whiz_tips[tip_index]

    def update_player(self, playerData):
        self.playerData = playerData

    def update_settings(self, gameSettings):
        self.gameSettings = gameSettings

    def perform_antics(self):
        antics = [
            "The farming whiz does a little dance.",
            "The farming whiz tells a funny joke.",
            "The farming whiz juggles some apples.",
            "The farming whiz takes a quick nap."
        ]
        antic_index = random.randint(0, len(antics) - 1)
        return antics[antic_index]
```