AI_Farmtopia/tests/test_farming_whiz.py
```python
import unittest
from AI_Farmtopia.farming_whiz import FarmingWhiz

class TestFarmingWhiz(unittest.TestCase):

    def setUp(self):
        self.farming_whiz = FarmingWhiz()

    def test_grant_advice(self):
        advice = self.farming_whiz.grant_advice()
        self.assertIsInstance(advice, str)

    def test_perform_antics(self):
        antics = self.farming_whiz.perform_antics()
        self.assertIsInstance(antics, str)

    def test_update_player_data(self):
        initial_player_data = self.farming_whiz.playerData
        self.farming_whiz.update_player_data({"level": 2, "experience": 150})
        updated_player_data = self.farming_whiz.playerData
        self.assertNotEqual(initial_player_data, updated_player_data)

    def test_update_game_settings(self):
        initial_game_settings = self.farming_whiz.gameSettings
        self.farming_whiz.update_game_settings({"difficulty": "hard"})
        updated_game_settings = self.farming_whiz.gameSettings
        self.assertNotEqual(initial_game_settings, updated_game_settings)

if __name__ == '__main__':
    unittest.main()
```